refactor: 에러 처리 및 리팩터링

+++ src/main/java/roomescape/reservationtime/service/ReservationTimeService.java
@@ -1,23 +1,27 @@
 package roomescape.reservationtime.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.reservation.dao.ReservationDao;
 import roomescape.reservationtime.dao.ReservationTimeDao;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.dto.ReservationTimeRequest;
 import roomescape.reservationtime.dto.ReservationTimeResponse;
+import roomescape.reservationtime.exception.ReservationTimeErrorCode;
+import roomescape.reservationtime.exception.ReservationTimeException;
 
 import java.util.List;
 
+@Transactional(readOnly = true)
 @Service
 public class ReservationTimeService {
 
-    @Autowired
-    private ReservationTimeDao reservationTimeDao;
+    private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     @Transactional
@@ -28,11 +32,24 @@ public class ReservationTimeService {
         return ReservationTimeResponse.from(createdTime);
     }
 
-    public List<ReservationTime> read() {
-        return reservationTimeDao.findAll();
+    public List<ReservationTimeResponse> read() {
+        List<ReservationTime> times = reservationTimeDao.findAll();
+
+        return times.stream()
+                .map(ReservationTimeResponse::from)
+                .toList();
     }
 
+    @Transactional
     public void delete(Long id) {
-        reservationTimeDao.delete(id);
+        if (reservationDao.existsByTimeId(id)) {
+            throw new ReservationTimeException(ReservationTimeErrorCode.RESERVATION_TIME_ALREADY_USED);
+        }
+
+        int deletedTimeCount = reservationTimeDao.delete(id);
+
+        if (deletedTimeCount == 0) {
+            throw new ReservationTimeException(ReservationTimeErrorCode.RESERVATION_TIME_NOT_FOUND);
+        }
     }
 }
